Move decimals computation to commons
